+++ packages/typespec-rust/CHANGELOG.md
@@ -18,6 +18,7 @@
 * Fixed an issue that could cause enum types to have invalid names.
 * Ensure that the local variable name for the `http::Request` doesn't collide with an existing parameter name.
 * Added missing reinjectable query parameters when creating the `Url` for a pageable's `next_link`.
+* Fixed incorrect code for enums that use numeric values instead of strings.
 
 ### Other Changes